+++ src/input.rs
@@ -1,13 +1,13 @@
 use std::io::{self, BufRead, BufReader};
 use std::fs::File;
 
-pub fn get_numeric_input(day: &str) -> Result<Vec<u32>, io::Error> {
+pub fn get_numeric_input(day: &str) -> Result<Vec<i64>, io::Error> {
     let mut output = Vec::new();
     let file_handle = File::open(format!("inputs/{}", day))?;
     let reader = BufReader::new(file_handle);
     for try_line in reader.lines() {
         match try_line {
-            Ok(line) => output.push(line.parse::<u32>().expect("invalid input format")),
+            Ok(line) => output.push(line.parse::<i64>().expect("invalid input format")),
             Err(e) => return Err(e)
         }
     }
